[TS migration] Migrate 'withNavigationFallback.js' HOC
